Add files via upload
Use 2023 healthcare stock excel for the training data in first cell for code to run. The last cell in code will be the prediction cell where you enter the 4 variables needed to make a prediction of [Buy] or [Skip] and the probabilities of the stock going up or down during the period of announcement date -> ex dividend date.